+++ src/PacketDataStream.h
@@ -54,7 +54,7 @@ public:
 			memcpy(&data[offset], d, len);
 			offset += len;
 		} else {
-			int l = left();
+			unsigned int l = left();
 			memset(&data[offset], 0, l);
 			offset += l;
 			overshoot += len - l;
@@ -99,7 +99,7 @@ public:
 
 	QByteArray dataBlock(quint32 len) {
 		if (len <= left()) {
-			QByteArray a(charPtr(), len);
+			QByteArray a(charPtr(), static_cast< int >(len));
 			offset += len;
 			return a;
 		} else {
@@ -109,7 +109,7 @@ public:
 	}
 
 protected:
-	void setup(unsigned char *d, int msize) {
+	void setup(unsigned char *d, unsigned int msize) {
 		data      = d;
 		offset    = 0;
 		overshoot = 0;
@@ -118,23 +118,24 @@ protected:
 	}
 
 public:
-	PacketDataStream(const unsigned char *d, int msize) { setup(const_cast< unsigned char * >(d), msize); };
+	PacketDataStream(const unsigned char *d, unsigned int msize) { setup(const_cast< unsigned char * >(d), msize); };
 
-	PacketDataStream(const char *d, int msize) {
+	PacketDataStream(const char *d, unsigned int msize) {
 		setup(const_cast< unsigned char * >(reinterpret_cast< const unsigned char * >(d)), msize);
 	};
 
-	PacketDataStream(char *d, int msize) { setup(reinterpret_cast< unsigned char * >(d), msize); };
+	PacketDataStream(char *d, unsigned int msize) { setup(reinterpret_cast< unsigned char * >(d), msize); };
 
-	PacketDataStream(unsigned char *d, int msize) { setup(d, msize); };
+	PacketDataStream(unsigned char *d, unsigned int msize) { setup(d, msize); };
 
 	PacketDataStream(const QByteArray &qba) {
-		setup(const_cast< unsigned char * >(reinterpret_cast< const unsigned char * >(qba.constData())), qba.size());
+		setup(const_cast< unsigned char * >(reinterpret_cast< const unsigned char * >(qba.constData())),
+			  static_cast< unsigned int >(qba.size()));
 	}
 
 	PacketDataStream(QByteArray &qba) {
 		unsigned char *ptr = reinterpret_cast< unsigned char * >(qba.data());
-		setup(ptr, qba.capacity());
+		setup(ptr, static_cast< unsigned int >(qba.capacity()));
 	}
 
 	PacketDataStream &operator<<(const quint64 value) {
@@ -236,7 +237,7 @@ public:
 
 	PacketDataStream &operator<<(const QByteArray &a) {
 		*this << a.size();
-		append(a.constData(), a.size());
+		append(a.constData(), static_cast< unsigned int >(a.size()));
 		return *this;
 	}
 
@@ -247,7 +248,7 @@ public:
 			len = left();
 			ok  = false;
 		}
-		a = QByteArray(reinterpret_cast< const char * >(&data[offset]), len);
+		a = QByteArray(reinterpret_cast< const char * >(&data[offset]), static_cast< int >(len));
 		offset += len;
 		return *this;
 	}
@@ -262,7 +263,7 @@ public:
 			len = left();
 			ok  = false;
 		}
-		s = QString::fromUtf8(reinterpret_cast< const char * >(&data[offset]), len);
+		s = QString::fromUtf8(reinterpret_cast< const char * >(&data[offset]), static_cast< int >(len));
 		offset += len;
 		return *this;
 	}
